Bug Fix: Unlink File If Target File Already Exists
Looks like the syscall rename deletes the target file if it already
exists. As far as I understood this 'hurts' the default behaviour of
the classic implementation that do 'link/unlink', since if the new
file already exists, link will fail.

So this commit implements 'rename' in a way that if the new file
already exists, it will be deleted. This way, other programs that
expect that 'rename' syscall exists could rely on it.