Removing restrict Keyword From Source Code
Although the _restrict_ keyword provides an hint to the compiler
which enables some sort of optimizations, it also could lead to
undefined behaviour if misused, if there are overlapping objects.

Moreover, since we are targeting self-hosting of Binutils/GCC,
the library could not use the _restrict_ keyword to it, since
a good amount of code in GCC is written in C++, which not
support restrict officially.

So the best approach here, is just get rid of all _restrict_
keywords, and thus, let the code more portable.