Enhancement: Multiples Data Program Header
In certain situations (such as C++ binaries) ELF files may contain more
than one data program header and in these cases Nanvix reads only the
first one found and crashes shortly afterwards when trying to load the
next ones.

To address this, support for multiples data headers (LOAD and RW) was
introduced in this commit. However, it is important to note that there
may be cases where there are multiple text headers, which is not
addressed in this commit (although easy to add). It is also important
to mention that Nanvix 'clears' only one BSS section (NOBITS) and if
there are more than one, they will not be cleaned, which may or may
not cause issues during execution.

In my tests, all binaries respected the following configuration: 1 text
program header (LOAD/RE), 2 data program headers (LOAD/RW) and only
one BSS section (NOBITS).