Enhancement: Preempt Process Only If Quantum Runs Out
Until now, Nanvix has always invoked the scheduler when executing
a system call, assuming that a syscall was time consuming enough
to disrupt other processes.

It turns out that one syscall does not consume that much time and
a single quantum may be sufficient to perform multiple syscalls.
This commit therefore changes the scheduler so that: 1) the
quantum is decremented even in kernel mode and 2) after performing
a syscall no preemption occurs, but returns to the userspace. If
the quantum runs out in kernel mode, the process will be scheduled
on the first clock interrupt _after_ starting to run in userspace,
however, I see no problem with that.