Bug Fix: Fixes issues when accessing more than 4MiB of heap
In Nanvix's current approach, when a heap region expanded to more than
4MiB, that is, a new page table, the system did not create this new
table and assumed it already exists, the same goes for miniregions.

This commit fixes the issue by adding an initial check for when the
page index is '0', and if so, create the page table and miniregion
as needed.