Enhancement: Process memory should be based on UMEM_SIZE
The total memory of a process should be based on userspace memory,
not total system memory, as total system memory also includes ramdisk,
kernel memory, kernel page pool, and so on. This commit, therefore,
makes the total memory of a process half of the maximum memory in
userspace.

However, note that this can lead to problems if more than one
process uses a lot of memory, as this does not take into account
the  _remaining_ amount of memory in userspace, but only the total.

A better approach would be to take into account the memory of all
processes rather than using a hardcoded value.