Bug Fix: Fix wrong pgtab mapping on demand-zero pages
When allocating a previously expanded user page as demand-zero, the
kernel bound the wrong PTEs in the page table, as a result of
incorrectly calculating the address inside the loop. The address
stored in 'addr', initially points to the last PTE to be allocated/
populated, and therefore, the following addresses must 'go up',
towards the top of the stack, not the other way around.

It took me 2 years to figure this out (and 10h (non-continuous)
debugging).